+++ packages/imperative/src/config/__tests__/ProfileInfo.TeamConfig.unit.test.ts
@@ -10,7 +10,6 @@
 */
 
 import * as fs from "fs";
-import * as os from "os";
 import * as path from "path";
 import * as jsonfile from "jsonfile";
 import * as lodash from "lodash";
@@ -25,7 +24,6 @@ import { ProfLocType } from "../src/doc/IProfLoc";
 import { IProfileSchema } from "../../profiles";
 import { AbstractSession, SessConstants } from "../../rest";
 import { ConfigAutoStore } from "../src/ConfigAutoStore";
-import { EnvironmentalVariableSettings } from "../../imperative/src/env/EnvironmentalVariableSettings";
 import { ImperativeConfig } from "../../utilities/src/ImperativeConfig";
 import { ImperativeError } from "../../error";
 import { IProfInfoUpdatePropOpts } from "../src/doc/IProfInfoUpdatePropOpts";
@@ -33,13 +31,14 @@ import { ConfigUtils } from "../src/ConfigUtils";
 import { ConfigProfiles } from "../src/api";
 import { IExtendersJsonOpts } from "../src/doc/IExtenderOpts";
 import { ConfigSchema } from "../src/ConfigSchema";
-import { Logger } from "../..";
+import { Logger } from "../../logger/src/Logger";
 
-jest.mock("../../events/src/ImperativeEventEmitter");
 
 const testAppNm = "ProfInfoApp";
 const testEnvPrefix = testAppNm.toUpperCase();
 const profileTypes = ["zosmf", "tso", "base", "dummy"];
+const testDir = path.join(__dirname, "__resources__");
+const teamProjDir = path.join(testDir, testAppNm + "_team_config_proj");
 
 function createNewProfInfo(newDir: string, opts?: IProfOpts): ProfileInfo {
     // create a new ProfileInfo in the desired directory
@@ -54,8 +53,6 @@ describe("TeamConfig ProfileInfo tests", () => {
     const tsoName = "tsoProfName";
     const tsoProfName = "LPAR1.tsoProfName";
     const tsoJsonLoc = "profiles.LPAR1.profiles." + tsoName;
-    const testDir = path.join(__dirname, "__resources__");
-    const teamProjDir = path.join(testDir, testAppNm + "_team_config_proj");
     const userTeamProjDir = path.join(testDir, testAppNm + "_user_and_team_config_proj");
     const teamHomeProjDir = path.join(testDir, testAppNm + "_home_team_config_proj");
     const largeTeamProjDir = path.join(testDir, testAppNm + "_large_team_config_proj");
@@ -151,65 +148,6 @@ describe("TeamConfig ProfileInfo tests", () => {
                     })
                 });
                 expect(ProfileInfo.onlyV1ProfilesExist).toBe(true);
-            });
-        });
-
-        describe("getZoweDir", () => {
-            const expectedLoadedConfig = {
-                name: "zowe",
-                defaultHome: path.join("z", "zowe"),
-                envVariablePrefix: "ZOWE"
-            };
-            let defaultHome: string;
-            let envReadSpy: any;
-            let homeDirSpy: any;
-            let loadedConfigOrig: any;
-
-            beforeAll(() => {
-                loadedConfigOrig = ImperativeConfig.instance.loadedConfig;
-            });
-
-            beforeEach(() => {
-                envReadSpy = jest.spyOn(EnvironmentalVariableSettings, "read").mockReturnValue({
-                    cliHome: { value: null }
-                } as any);
-                homeDirSpy = jest.spyOn(os, "homedir").mockReturnValue(expectedLoadedConfig.defaultHome);
-                ImperativeConfig.instance.loadedConfig = undefined as any;
-                defaultHome = path.join(expectedLoadedConfig.defaultHome, ".zowe");
-            });
-
-            afterAll(() => {
-                ImperativeConfig.instance.loadedConfig = loadedConfigOrig;
-                envReadSpy.mockRestore();
-                homeDirSpy.mockRestore();
-            });
-
-            it("should return the ENV cliHome even if loadedConfig is set in the process", () => {
-                jest.spyOn(EnvironmentalVariableSettings, "read").mockReturnValue({ cliHome: { value: "test" } } as any);
-                expect(ImperativeConfig.instance.loadedConfig).toBeUndefined();
-                expect(ProfileInfo.getZoweDir()).toEqual("test");
-                expect(ImperativeConfig.instance.loadedConfig).toEqual({ ...expectedLoadedConfig, defaultHome });
-            });
-
-            it("should return the defaultHome and set loadedConfig if undefined", () => {
-                expect(ImperativeConfig.instance.loadedConfig).toBeUndefined();
-                expect(ProfileInfo.getZoweDir()).toEqual(defaultHome);
-                expect(ImperativeConfig.instance.loadedConfig).toEqual({ ...expectedLoadedConfig, defaultHome });
-            });
-
-            it("should return the defaultHome and reset loadedConfig if defaultHome changes", () => {
-                expect(ImperativeConfig.instance.loadedConfig).toBeUndefined();
-                ImperativeConfig.instance.loadedConfig = { ...expectedLoadedConfig, defaultHome: "test" };
-                expect(ImperativeConfig.instance.loadedConfig?.defaultHome).toEqual("test");
-                expect(ProfileInfo.getZoweDir()).toEqual(defaultHome);
-                expect(ImperativeConfig.instance.loadedConfig).toEqual({ ...expectedLoadedConfig, defaultHome });
-            });
-
-            it("should return the defaultHome without resetting loadedConfig", () => {
-                expect(ImperativeConfig.instance.loadedConfig).toBeUndefined();
-                ImperativeConfig.instance.loadedConfig = expectedLoadedConfig;
-                expect(ProfileInfo.getZoweDir()).toEqual(defaultHome);
-                expect(ImperativeConfig.instance.loadedConfig).toEqual({ ...expectedLoadedConfig, defaultHome });
             });
         });
 
@@ -542,6 +480,9 @@ describe("TeamConfig ProfileInfo tests", () => {
     });
 
     describe("mergeArgsForProfile", () => {
+        beforeEach(() => {
+            (ImperativeConfig as any).mInstance = null;
+        });
         afterEach(() => {
             delete process.env[envHost];
             delete process.env[envPort];
@@ -919,6 +860,7 @@ describe("TeamConfig ProfileInfo tests", () => {
         it("should not look for secure properties in the global-layer base profile if it does not exist", async () => {
             process.env[testEnvPrefix + "_CLI_HOME"] = nestedTeamProjDir;
             const profInfo = createNewProfInfo(userTeamProjDir);
+            process.env[testEnvPrefix + "_CLI_HOME"] = nestedTeamProjDir;
             await profInfo.readProfilesFromDisk();
             const profiles = profInfo.getAllProfiles();
             const desiredProfile = "TEST001.first";
@@ -933,7 +875,7 @@ describe("TeamConfig ProfileInfo tests", () => {
                 }
                 expect(unexpectedError).toBeUndefined();
             } else {
-                expect("Profile " + desiredProfile + "not found").toBeUndefined();
+                expect("Profile " + desiredProfile + " not found").toBeUndefined();
             }
             expect(mergedArgs.missingArgs.find(a => a.argName === "user")?.secure).toBeTruthy();
             expect(mergedArgs.missingArgs.find(a => a.argName === "password")?.secure).toBeTruthy();
@@ -1213,7 +1155,7 @@ describe("TeamConfig ProfileInfo tests", () => {
             const storageSpy = jest.spyOn(ConfigAutoStore as any, "_storeSessCfgProps").mockResolvedValue(undefined);
             const profiles = [{ type: "test", schema: {} as any }];
             ImperativeConfig.instance.loadedConfig.profiles = profiles;
-            ImperativeConfig.instance.loadedConfig.baseProfile = null;
+            ImperativeConfig.instance.loadedConfig.baseProfile = undefined;
 
             let caughtError;
             try {
@@ -1343,11 +1285,11 @@ describe("TeamConfig ProfileInfo tests", () => {
             const mergedArgs = profInfo.mergeArgsForProfile(profAttrs);
 
             const userArg = mergedArgs.knownArgs.find((arg) => arg.argName === "user");
-            expect(userArg.argValue).toBe("userNameBase");
+            expect(userArg?.argValue).toBe("userNameBase");
             expect(profInfo.loadSecureArg(userArg as IProfArgAttrs)).toBe("userNameBase");
 
             const passwordArg = mergedArgs.knownArgs.find((arg) => arg.argName === "password");
-            expect(passwordArg.argValue).toBe("passwordBase");
+            expect(passwordArg?.argValue).toBe("passwordBase");
             expect(profInfo.loadSecureArg(passwordArg as IProfArgAttrs)).toBe("passwordBase");
         });
 
@@ -1358,10 +1300,10 @@ describe("TeamConfig ProfileInfo tests", () => {
             const mergedArgs = profInfo.mergeArgsForProfile(profAttrs, { getSecureVals: true });
 
             const userArg = mergedArgs.knownArgs.find((arg) => arg.argName === "user");
-            expect(userArg.argValue).toBe("userNameBase");
+            expect(userArg?.argValue).toBe("userNameBase");
 
             const passwordArg = mergedArgs.knownArgs.find((arg) => arg.argName === "password");
-            expect(passwordArg.argValue).toBe("passwordBase");
+            expect(passwordArg?.argValue).toBe("passwordBase");
         });
 
         it("should treat secure arg as plain text if loaded from environment variable", async () => {
@@ -1388,7 +1330,7 @@ describe("TeamConfig ProfileInfo tests", () => {
                 profInfo.loadSecureArg({
                     argName: "test",
                     dataType: "string",
-                    argValue: undefined,
+                    argValue: undefined as any,
                     argLoc: { locType: ProfLocType.DEFAULT }
                 });
             } catch (error) {
@@ -1415,8 +1357,12 @@ describe("TeamConfig ProfileInfo tests", () => {
             const profAttrs = profInfo.getDefaultProfile("zosmf") as IProfAttrs;
             const osLocInfo = profInfo.getOsLocInfo(profAttrs);
             const expectedObjs = [
-                { name: profAttrs.profName, path: profAttrs.profLoc.osLoc[0], user: false, global: false },
-                { name: profAttrs.profName, path: profAttrs.profLoc.osLoc[0], user: false, global: true }
+                { name: profAttrs.profName, path: profAttrs.profLoc.osLoc?.[0], user: false, global: false },
+                // TODO(zFernand0): Investigate why only the team project is present in the osLoc array
+                //      Possible reason: global layer not loaded by getAllProfiles()
+                //      ----
+                //      Reseting the loaded configuration in `getZoweDir` may be the root cause.
+                { name: profAttrs.profName, path: profAttrs.profLoc.osLoc?.[0], user: false, global: true }
             ];
             expect(osLocInfo).toBeDefined();
             expect(osLocInfo.length).toBe(expectedObjs.length);
@@ -1468,58 +1414,6 @@ describe("TeamConfig ProfileInfo tests", () => {
         };
 
         // begin schema management tests
-        describe("readExtendersJsonFromDisk", () => {
-            // case 1: the JSON file doesn't exist at time of read
-            it("writes an empty extenders.json file if it doesn't exist on disk", async () => {
-                const profInfo = createNewProfInfo(teamProjDir);
-                (profInfo as any).mExtendersJson = { profileTypes: {} };
-                jest.spyOn(fs, "existsSync").mockReturnValueOnce(false);
-                ProfileInfo.readExtendersJsonFromDisk();
-                expect(writeFileSyncMock).toHaveBeenCalled();
-            });
-
-            // case 2: JSON file exists on-disk at time of read
-            it("reads extenders.json from disk if it exists", async () => {
-                const readFileSyncMock = jest.spyOn(jsonfile, "readFileSync").mockReturnValueOnce({ profileTypes: {
-                    "test": {
-                        from: ["Zowe Client App"]
-                    }
-                } });
-                const profInfo = createNewProfInfo(teamProjDir);
-                jest.spyOn(fs, "existsSync").mockReturnValueOnce(true);
-                (profInfo as any).mExtendersJson = ProfileInfo.readExtendersJsonFromDisk();
-                expect(readFileSyncMock).toHaveBeenCalled();
-                expect((profInfo as any).mExtendersJson).toEqual({
-                    profileTypes: {
-                        "test": {
-                            from: ["Zowe Client App"]
-                        }
-                    }
-                });
-            });
-        });
-
-        describe("writeExtendersJson", () => {
-            // case 1: Write operation is successful
-            it("returns true if written to disk successfully", async () => {
-                const profInfo = createNewProfInfo(teamProjDir);
-                (profInfo as any).mExtendersJson = { profileTypes: {} };
-                await profInfo.readProfilesFromDisk({ homeDir: teamHomeProjDir });
-                expect(ProfileInfo.writeExtendersJson((profInfo as any).mExtendersJson)).toBe(true);
-                expect(writeFileSyncMock).toHaveBeenCalled();
-            });
-
-            // case 2: Write operation is unsuccessful
-            it("returns false if it couldn't write to disk", async () => {
-                const profInfo = createNewProfInfo(teamProjDir);
-                (profInfo as any).mExtendersJson = { profileTypes: {} };
-                await profInfo.readProfilesFromDisk({ homeDir: teamHomeProjDir });
-                writeFileSyncMock.mockImplementation(() => { throw new Error(); });
-                expect(ProfileInfo.writeExtendersJson((profInfo as any).mExtendersJson)).toBe(false);
-                expect(writeFileSyncMock).toHaveBeenCalled();
-            });
-        });
-
         describe("updateSchemaAtLayer", () => {
             const getBlockMocks = () => {
                 return {
@@ -1671,7 +1565,7 @@ describe("TeamConfig ProfileInfo tests", () => {
                 } as any);
                 const updateSchemaAtLayerMock = jest.spyOn((ProfileInfo as any).prototype, "updateSchemaAtLayer")
                     .mockReturnValue(expected.res.success);
-                const writeExtendersJsonMock = jest.spyOn(ProfileInfo, "writeExtendersJson").mockImplementation();
+                const writeExtendersJsonMock = jest.spyOn(ConfigUtils, "writeExtendersJson").mockImplementation();
                 const res = profInfo.addProfileTypeToSchema("some-type", { ...testCase, sourceApp: "Zowe Client App" });
                 if (expected.res.success) {
                     expect(updateSchemaAtLayerMock).toHaveBeenCalled();