+++ packages/imperative/CHANGELOG.md
@@ -7,6 +7,16 @@ All notable changes to the Imperative package will be documented in this file.
 - BugFix: Resolved bug that resulted in each plug-in to have identical public registries regardless of actual installation location/reference. [#2189](https://github.com/zowe/zowe-cli/pull/2189)
 - BugFix: Resolved bug that resulted in every plug-in to have the same registry location field as the first if multiple plugins were installed in the same command. [#2189](https://github.com/zowe/zowe-cli/pull/2189)
 
+## `8.0.0-next.202407112150`
+
+- Enhancement: Add client-side custom-event handling capabilities. [#2136](https://github.com/zowe/zowe-cli/pull/2136)
+- Next-Breaking: Refactored the Imperative Event Emitter class. [#2136](https://github.com/zowe/zowe-cli/pull/2136)
+  - Removed the `ImperativeEventEmitter` class.
+  - Added an `EventProcessor` class to handle event listening and emitting.
+  - Added an `EventOperator` class to handle creation and deletion of `EventProcessors`.
+  - Added an `EventUtils` class to contain all common utility methods for the Client Event Handling capabilities.
+  - Added `IEmitter`, `IWatcher`, and `IEmitterAndWatcher` interfaces to expose what application developers should see.
+
 ## `8.0.0-next.202407051717`
 
 - BugFix: V3 Breaking: Modified the ConvertV1Profiles.convert API to accept a new ProfileInfo option and initialize components sufficiently to enable VSCode apps to convert V1 profiles. [#2170](https://github.com/zowe/zowe-cli/issues/2170)
@@ -86,7 +96,7 @@ All notable changes to the Imperative package will be documented in this file.
 
 ## `8.0.0-next.202405151329`
 
-- Enhancement: Add client-side event handling capabilities. [#1987](https://github.com/zowe/zowe-cli/pull/1987)
+- Enhancement: Add client-side event handling capabilities. [#1987](https://github.com/zowe/zowe-cli/issues/1987)
 
 ## `8.0.0-next.202405061946`